Revert "Added Baseflow plugin template"
This reverts commit b428726ee38a6b5fcff21d82eaf5f4ecd2b63b30.
