+++ permission_handler/example/lib/main.dart
@@ -1,158 +1,127 @@
+import 'dart:io';
+
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:permission_handler/permission_handler.dart';
-import 'package:permission_handler_example/template/globals.dart';
 
-void main() {
-  runApp(BaseflowPluginExample());
-}
+void main() => runApp(MyApp());
 
-class BaseflowPluginExample extends StatelessWidget {
-  final MaterialColor themeMaterialColor =
-      createMaterialColor(const Color.fromRGBO(48, 49, 60, 1));
-
+/// Example Flutter Application demonstrating the functionality of the
+/// Permission Handler plugin.
+class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Baseflow ${Globals.pluginName}',
-      theme: ThemeData(
-        accentColor: Colors.white60,
-        backgroundColor: const Color.fromRGBO(48, 49, 60, 0.8),
-        buttonTheme: ButtonThemeData(
-          buttonColor: themeMaterialColor.shade500,
-          disabledColor: themeMaterialColor.withRed(200),
-          splashColor: themeMaterialColor.shade50,
-          textTheme: ButtonTextTheme.primary,
+      home: Scaffold(
+        appBar: AppBar(
+          title: const Text('Plugin example app'),
+          actions: <Widget>[
+            IconButton(
+              icon: const Icon(Icons.settings),
+              onPressed: () async {
+                var hasOpened = openAppSettings();
+                debugPrint('App Settings opened: ' + hasOpened.toString());
+              },
+            )
+          ],
         ),
-        bottomAppBarColor: const Color.fromRGBO(57, 58, 71, 1),
-        hintColor: themeMaterialColor.shade500,
-        primarySwatch: createMaterialColor(const Color.fromRGBO(48, 49, 60, 1)),
-        textTheme: TextTheme(
-          bodyText1: TextStyle(
-            color: Colors.white,
-            fontSize: 16,
-            height: 1.3,
-          ),
-          bodyText2: TextStyle(
-            color: Colors.white,
-            fontSize: 18,
-            height: 1.2,
-          ),
-          button: TextStyle(color: Colors.white),
-          headline1: TextStyle(
-            color: Colors.white,
-            fontSize: 18,
-          ),
-        ),
-        visualDensity: VisualDensity.adaptivePlatformDensity,
-        inputDecorationTheme: const InputDecorationTheme(
-          fillColor: Color.fromRGBO(37, 37, 37, 1),
-          filled: true,
+        body: Center(
+          child: ListView(
+              children: Permission.values
+                  .where((Permission permission) {
+                    if (Platform.isIOS) {
+                      return permission != Permission.unknown &&
+                          permission != Permission.sms &&
+                          //permission != Permission.storage &&
+                          permission != Permission.ignoreBatteryOptimizations &&
+                          permission != Permission.accessMediaLocation;
+                    } else {
+                      return permission != Permission.unknown &&
+                          permission != Permission.mediaLibrary &&
+                          permission != Permission.photos &&
+                          permission != Permission.reminders;
+                    }
+                  })
+                  .map((permission) => PermissionWidget(permission))
+                  .toList()),
         ),
       ),
-      home: MyHomePage(title: 'Baseflow ${Globals.pluginName} example app'),
     );
   }
-
-  static MaterialColor createMaterialColor(Color color) {
-    List strengths = <double>[.05];
-    Map swatch = <int, Color>{};
-    final int r = color.red, g = color.green, b = color.blue;
-
-    for (int i = 1; i < 10; i++) {
-      strengths.add(0.1 * i);
-    }
-    strengths.forEach((strength) {
-      final double ds = 0.5 - strength;
-      swatch[(strength * 1000).round()] = Color.fromRGBO(
-        r + ((ds < 0 ? r : (255 - r)) * ds).round(),
-        g + ((ds < 0 ? g : (255 - g)) * ds).round(),
-        b + ((ds < 0 ? b : (255 - b)) * ds).round(),
-        1,
-      );
-    });
-    return MaterialColor(color.value, swatch);
-  }
 }
 
-class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+/// Permission widget which displays a permission and allows users to request
+/// the permissions.
+class PermissionWidget extends StatefulWidget {
+  /// Constructs a [PermissionWidget] for the supplied [Permission].
+  const PermissionWidget(this._permission);
 
-  final String title;
+  final Permission _permission;
 
   @override
-  _MyHomePageState createState() => _MyHomePageState();
+  _PermissionState createState() => _PermissionState(_permission);
 }
 
-class _MyHomePageState extends State<MyHomePage> {
-  static final PageController _pageController = PageController(initialPage: 0);
-  int currentPage = 0;
+class _PermissionState extends State<PermissionWidget> {
+  _PermissionState(this._permission);
+
+  final Permission _permission;
+  PermissionStatus _permissionStatus = PermissionStatus.undetermined;
+
+  @override
+  void initState() {
+    super.initState();
+
+    _listenForPermissionStatus();
+  }
+
+  void _listenForPermissionStatus() async {
+    final status = await _permission.status;
+    setState(() => _permissionStatus = status);
+  }
+
+  Color getPermissionColor() {
+    switch (_permissionStatus) {
+      case PermissionStatus.denied:
+        return Colors.red;
+      case PermissionStatus.granted:
+        return Colors.green;
+      default:
+        return Colors.grey;
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Theme.of(context).bottomAppBarColor,
-        title: Center(
-          child: Image.asset(
-            'res/images/baseflow_logo_def_light-02.png',
-            width: 140,
-          ),
-        ),
-        actions: <Widget>[
-          IconButton(
-            icon: const Icon(Icons.settings),
-            color: Colors.white,
-            onPressed: () async {
-              var hasOpened = openAppSettings();
-              debugPrint('App Settings opened: ' + hasOpened.toString());
-            },
-          )
-        ],
+    return ListTile(
+      title: Text(_permission.toString()),
+      subtitle: Text(
+        _permissionStatus.toString(),
+        style: TextStyle(color: getPermissionColor()),
       ),
-      backgroundColor: Theme.of(context).backgroundColor,
-      body: PageView(
-        controller: _pageController,
-        children: Globals.pages,
-        onPageChanged: (page) {
-          setState(() {
-            currentPage = page;
-          });
-        },
-      ),
-      bottomNavigationBar: _bottomAppBar(),
+      trailing: IconButton(
+          icon: const Icon(Icons.info),
+          onPressed: () {
+            checkServiceStatus(context, _permission);
+          }),
+      onTap: () {
+        requestPermission(_permission);
+      },
     );
   }
 
-  BottomAppBar _bottomAppBar() {
-    return BottomAppBar(
-      elevation: 5,
-      color: Theme.of(context).bottomAppBarColor,
-      child: Row(
-        mainAxisSize: MainAxisSize.max,
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: List.unmodifiable(() sync* {
-          for (int i = 0; i < Globals.pages.length; i++) {
-            yield Expanded(
-              child: IconButton(
-                iconSize: 30,
-                icon: Icon(Globals.icons.elementAt(i)),
-                color: _bottomAppBarIconColor(i),
-                onPressed: () => _animateToPage(i),
-              ),
-            );
-          }
-        }()),
-      ),
-    );
+  void checkServiceStatus(BuildContext context, Permission permission) async {
+    Scaffold.of(context).showSnackBar(SnackBar(
+      content: Text((await permission.status).toString()),
+    ));
   }
 
-  void _animateToPage(int page) {
-    _pageController.animateToPage(page,
-        duration: const Duration(milliseconds: 200), curve: Curves.linear);
-  }
+  Future<void> requestPermission(Permission permission) async {
+    final status = await permission.request();
 
-  Color _bottomAppBarIconColor(int page) {
-    return currentPage == page ? Colors.white : Theme.of(context).accentColor;
+    setState(() {
+      print(status);
+      _permissionStatus = status;
+      print(_permissionStatus);
+    });
   }
 }
