--- permission_handler/example/lib/plugin_example/permission_widget.dart
@@ -1,85 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:permission_handler/permission_handler.dart';
-
-/// Permission widget which displays a permission and allows users to request
-/// the permissions.
-class PermissionWidget extends StatefulWidget {
-  /// Constructs a [PermissionWidget] for the supplied [Permission].
-  const PermissionWidget(this._permission);
-
-  final Permission _permission;
-
-  @override
-  _PermissionState createState() => _PermissionState(_permission);
-}
-
-class _PermissionState extends State<PermissionWidget> {
-  _PermissionState(this._permission);
-
-  final Permission _permission;
-  PermissionStatus _permissionStatus = PermissionStatus.undetermined;
-
-  @override
-  void initState() {
-    super.initState();
-
-    _listenForPermissionStatus();
-  }
-
-  void _listenForPermissionStatus() async {
-    final status = await _permission.status;
-    setState(() => _permissionStatus = status);
-  }
-
-  Color getPermissionColor() {
-    switch (_permissionStatus) {
-      case PermissionStatus.denied:
-        return Colors.red;
-      case PermissionStatus.granted:
-        return Colors.green;
-      default:
-        return Colors.grey;
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return ListTile(
-      title: Text(
-        _permission.toString(),
-        style: Theme.of(context).textTheme.bodyText1,
-      ),
-      subtitle: Text(
-        _permissionStatus.toString(),
-        style: TextStyle(color: getPermissionColor()),
-      ),
-      trailing: IconButton(
-          icon: const Icon(
-            Icons.info,
-            color: Colors.white,
-          ),
-          onPressed: () {
-            checkServiceStatus(context, _permission);
-          }),
-      onTap: () {
-        requestPermission(_permission);
-      },
-    );
-  }
-
-  void checkServiceStatus(BuildContext context, Permission permission) async {
-    Scaffold.of(context).showSnackBar(SnackBar(
-      content: Text((await permission.status).toString()),
-    ));
-  }
-
-  Future<void> requestPermission(Permission permission) async {
-    final status = await permission.request();
-
-    setState(() {
-      print(status);
-      _permissionStatus = status;
-      print(_permissionStatus);
-    });
-  }
-}
