+++ permission_handler/example/pubspec.yaml
@@ -15,12 +15,5 @@ dev_dependencies:
   permission_handler:
     path: ../
 
-  url_launcher: ^5.4.11
-  cupertino_icons: ^0.1.3
-
 flutter:
   uses-material-design: true
-
-  assets:
-    - res/images/baseflow_logo_def_light-02.png
-    - res/images/poweredByBaseflowLogoLight@3x.png
